Rename the helper
