+++ validator/tools/ghHelper.js
